requirements.txt
streamlit==1.44.1
sqlalchemy==2.0.38
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
APScheduler==3.11.0
requests==2.32.3
httpx==0.28.1
python-dotenv==1.0.1
plotly==6.0.1
pytest==8.3.5
